balance: add new balance method and make station activation explicit
A new, additional balance method has been added to balance the line by
shortest station time. However, this method requires that we explicitly
make a station active instead of treating it as active if it has at least
one task assignment. The other balance method did not care about active/
inactive, so this functionality was not needed.

The terminating case for the new balance method also depends on the line
be paced and that global work does not exceed global capacity. These
line constraints are done in the ValidateParams() method. I would prefer
that these "global" line constraints be broken out into more formal
definitions like we currently have for task-station assignment constraints.